requirements.txt
PyMuPDF 
faiss-cpu 
sentence-transformers
requests